+++ src/com/reco1l/ibancho/RoomAPI.kt
@@ -1,9 +1,17 @@
-package com.reco1l.api.ibancho
+package com.reco1l.ibancho
 
 import com.dgsrz.bancho.security.SecurityUtils
-import com.reco1l.api.ibancho.data.*
-import com.reco1l.api.ibancho.data.Room
-import com.reco1l.api.ibancho.data.RoomTeam
+import com.reco1l.ibancho.data.Room
+import com.reco1l.ibancho.data.PlayerStatus
+import com.reco1l.ibancho.data.RoomStatus
+import com.reco1l.ibancho.data.RoomTeam
+import com.reco1l.ibancho.data.TeamMode
+import com.reco1l.ibancho.data.WinCondition
+import com.reco1l.ibancho.data.parseBeatmap
+import com.reco1l.ibancho.data.parseGameplaySettings
+import com.reco1l.ibancho.data.parseMods
+import com.reco1l.ibancho.data.parsePlayer
+import com.reco1l.ibancho.data.parsePlayers
 import com.reco1l.osu.multiplayer.Multiplayer
 import io.socket.client.IO
 import io.socket.client.Socket
