+++ src/com/reco1l/ibancho/data/Room.kt
@@ -1,4 +1,4 @@
-package com.reco1l.api.ibancho.data
+package com.reco1l.ibancho.data
 
 import com.reco1l.osu.multiplayer.Multiplayer
