+++ src/com/reco1l/ibancho/data/RoomGameplaySettings.kt
@@ -1,4 +1,4 @@
-package com.reco1l.api.ibancho.data
+package com.reco1l.ibancho.data
 
 /**
  * Represents gameplay-related settings of a multiplayer room.
