+++ src/com/reco1l/ibancho/data/RoomMods.kt
@@ -1,4 +1,4 @@
-package com.reco1l.api.ibancho.data
+package com.reco1l.ibancho.data
 
 import com.reco1l.osu.conversion.modsToReadable
 import com.reco1l.osu.multiplayer.Multiplayer
