+++ src/com/reco1l/osu/multiplayer/LobbyRoomList.kt
@@ -3,12 +3,12 @@ package com.reco1l.osu.multiplayer
 import android.app.AlertDialog
 import android.view.inputmethod.EditorInfo
 import android.widget.EditText
-import com.reco1l.api.ibancho.RoomAPI
-import com.reco1l.api.ibancho.data.Room
-import com.reco1l.api.ibancho.data.RoomStatus.*
-import com.reco1l.api.ibancho.data.TeamMode.HEAD_TO_HEAD
-import com.reco1l.api.ibancho.data.TeamMode.TEAM_VS_TEAM
-import com.reco1l.api.ibancho.data.WinCondition.*
+import com.reco1l.ibancho.RoomAPI
+import com.reco1l.ibancho.data.Room
+import com.reco1l.ibancho.data.RoomStatus.*
+import com.reco1l.ibancho.data.TeamMode.HEAD_TO_HEAD
+import com.reco1l.ibancho.data.TeamMode.TEAM_VS_TEAM
+import com.reco1l.ibancho.data.WinCondition.*
 import com.reco1l.osu.mainThread
 import com.reco1l.osu.ui.ScrollableList
 import com.reco1l.toolkt.kotlin.async
