+++ src/com/reco1l/osu/multiplayer/LobbyScene.kt
@@ -3,8 +3,8 @@ package com.reco1l.osu.multiplayer
 import android.net.Uri
 import android.util.Log
 import com.dgsrz.bancho.security.SecurityUtils
-import com.reco1l.api.ibancho.LobbyAPI
-import com.reco1l.api.ibancho.RoomAPI
+import com.reco1l.ibancho.LobbyAPI
+import com.reco1l.ibancho.RoomAPI
 import com.reco1l.osu.updateThread
 import com.reco1l.toolkt.kotlin.async
 import org.anddev.andengine.entity.modifier.LoopEntityModifier
