+++ src/com/reco1l/osu/multiplayer/Multiplayer.kt
@@ -2,9 +2,9 @@ package com.reco1l.osu.multiplayer
 
 import android.text.format.DateFormat
 import android.util.Log
-import com.reco1l.api.ibancho.RoomAPI
-import com.reco1l.api.ibancho.data.Room
-import com.reco1l.api.ibancho.data.RoomPlayer
+import com.reco1l.ibancho.RoomAPI
+import com.reco1l.ibancho.data.Room
+import com.reco1l.ibancho.data.RoomPlayer
 import com.reco1l.osu.conversion.jsonToScoreboardItem
 import com.reco1l.osu.conversion.jsonToStatistic
 import com.reco1l.toolkt.kotlin.formatTimeMilliseconds
