+++ src/com/reco1l/osu/multiplayer/RoomPlayerList.kt
@@ -1,8 +1,10 @@
 package com.reco1l.osu.multiplayer
 
-import com.reco1l.api.ibancho.data.*
-import com.reco1l.api.ibancho.data.PlayerStatus.*
-import com.reco1l.api.ibancho.data.RoomTeam.*
+import com.reco1l.ibancho.data.PlayerStatus.*
+import com.reco1l.ibancho.data.Room
+import com.reco1l.ibancho.data.RoomPlayer
+import com.reco1l.ibancho.data.RoomTeam.*
+import com.reco1l.ibancho.data.TeamMode
 import com.reco1l.osu.ui.ScrollableList
 import org.anddev.andengine.entity.primitive.Rectangle
 import org.anddev.andengine.entity.sprite.Sprite
