+++ src/com/reco1l/osu/multiplayer/RoomPlayerMenu.kt
@@ -9,8 +9,8 @@ import com.edlplan.ui.BaseAnimationListener
 import com.edlplan.ui.EasingHelper
 import com.edlplan.ui.fragment.BaseFragment
 import com.edlplan.ui.fragment.WebViewFragment
-import com.reco1l.api.ibancho.RoomAPI
-import com.reco1l.api.ibancho.data.RoomPlayer
+import com.reco1l.ibancho.RoomAPI
+import com.reco1l.ibancho.data.RoomPlayer
 import ru.nsu.ccfit.zuev.osuplus.R
 import ru.nsu.ccfit.zuev.osu.GlobalManager.getInstance as getGlobal
