+++ src/ru/nsu/ccfit/zuev/osu/scoring/StatisticV2.java
@@ -7,7 +7,7 @@ import java.util.EnumSet;
 import java.util.Locale;
 import java.util.Random;
 
-import com.reco1l.api.ibancho.data.WinCondition;
+import com.reco1l.ibancho.data.WinCondition;
 import com.reco1l.osu.multiplayer.Multiplayer;
 import org.jetbrains.annotations.Nullable;
 import org.json.JSONObject;
